bugfix: don't require all fields to have changed before submit
An earlier commit introduced using redux-form handleSubmit in the
personal data form. A side effect of that was that validation was done
before submit too, not just on changes.

When validation was done per-field, it was apparently an error to have
an un-changed value, but when validation is done for the whole form this
isn't true anymore.

Anyways, it seems it was possible to replace something like 40 lines of
code with 'disabled={loading || pristine || invalid || submitting}'.