handle browser auto-fill as best as we can
When Chrome auto-fills username and password, Javascript doesn't see the
values until the user clicks somewhere so we can't enable the login
button based on form validators that require a value.

It seems the best we can do is make the "required" a submitError, with a
check when the user clicks "Log in". This avoids all the confusion with
the user seeing a value on the screen while the Javascript does not, at
the relatively small cost of not giving an error message about a missing
value until the user has clicked the button once.