remove group management
This code has been sitting around for a long time without being
released.

It broke with new dependencies, and it seems more probable that we will
re-implement it from scratch whenever we want to add groups to eduID.

We can still dig it up from the git history in the future if we want to
make an attempt to get it working.